fix(codec): internal fails (#498)

+++ libs/jwst-codec/src/doc/types/map.rs
@@ -66,8 +66,14 @@ pub(crate) trait MapType: AsInner<Inner = YTypeRef> {
     fn remove(&mut self, key: impl AsRef<str>) {
         let mut inner = self.as_inner().get().unwrap().write().unwrap();
         let node = inner.map.as_ref().and_then(|map| map.get(key.as_ref()));
-        if let Some(item) = ItemRef::from(node).get() {
-            DocStore::delete_item(item, Some(&mut inner));
+        if let Some(store) = inner.store.upgrade() {
+            let mut store = store.write().unwrap();
+            if let Some(item) = ItemRef::from(node).get() {
+                store
+                    .delete_set
+                    .add(item.id.client, item.id.clock, item.len());
+                DocStore::delete_item(item, Some(&mut inner));
+            }
         }
     }
